shuffle common/rep_interchange code for mvp
